requirements.txt
python-telegram-bot==13.4.1

# for .env file support
python-dotenv